move giflib library files
