+++ Source/UEgiflib/Private/giflib/dgif_lib.c
