+++ Source/UEgiflib/Private/giflib/egif_lib.c
