+++ Source/UEgiflib/Private/giflib/gif_err.c
