+++ Source/UEgiflib/Private/giflib/gif_font.c
