+++ Source/UEgiflib/Private/giflib/gif_lib.h
