+++ Source/UEgiflib/Private/giflib/gif_lib_private.h
