+++ Source/UEgiflib/Private/giflib/gifalloc.c
