+++ Source/UEgiflib/Private/giflib/quantize.c
